fix: resolve all pre-commit violations and pre-existing errors
