+++ acro/table_utils.py
@@ -71,7 +71,7 @@ def drop_duplicate_columns(outcome: pd.DataFrame) -> pd.DataFrame:
     return outcome
 
 
-def collate_risk_assessments(  # noqa: C901, PLR0912
+def collate_risk_assessments(
     table: DataFrame, allcheckresults: dict[str, ChecksResults]
 ) -> DataFrame:
     """Collate the Risk Assessment for a table.
