+++ foo.txt
@@ -1,0 +1,1 @@
+Your text goes here
